+++ core/headers/NeuromorphicTimeSpace/SlabToFabricConverterAndCordinator.h
@@ -37,8 +37,8 @@ namespace PredictedAdaptedEncoding
         void ShutDownFabric() noexcept;
 
         bool InitializeFabric(
-            uint16_t slot_count,
-            size_t slot_cell_count = MINIMUM_APC_CELL_COUNT,
+            uint32_t slot_count,
+            uint32_t slot_cell_count = MINIMUM_APC_CELL_COUNT,
             uint8_t slab_id = APCDataStructure::BRANCH_VERSION,
             uint32_t fabric_thread_capacity = CoreOfFabricCoordinator::DEFAULT_THREAD_TABLE_CAPACITY
         ) noexcept;
